@@ -3,23 +3,22 @@
   "version": "1.0.0",
   "private": true,
   "dependencies": {
-    "@emotion/react": "^11.8.1",
+    "@emotion/react": "^11.9.0",
     "@emotion/styled": "^11.8.1",
-    "@mui/icons-material": "^5.4.4",
-    "@mui/material": "^5.4.4",
-    "@mui/styles": "^5.6.2",
+    "@mui/icons-material": "^5.8.2",
+    "@mui/material": "^5.8.2",
     "@testing-library/jest-dom": "^5.16.2",
     "@testing-library/react": "^12.1.3",
     "@testing-library/user-event": "^13.5.0",
-    "i18next-browser-languagedetector": "^6.1.3",
-    "i18next-http-backend": "^1.3.2",
-    "react": "^17.0.2",
-    "react-device-detect": "^2.1.2",
-    "react-dom": "^17.0.2",
+    "i18next-browser-languagedetector": "^6.1.4",
+    "i18next-http-backend": "^1.4.1",
+    "react": "^18.1.0",
+    "react-device-detect": "^2.2.2",
+    "react-dom": "^18.1.0",
     "react-helmet-async": "^1.3.0",
-    "react-i18next": "^11.16.2",
-    "react-router-dom": "^6.2.2",
-    "react-scripts": "5.0.0",
+    "react-i18next": "^11.17.0",
+    "react-router-dom": "^6.3.0",
+    "react-scripts": "^5.0.1",
     "react-swipeable": "^7.0.0",
     "react-transition-group": "^4.4.2",
     "web-vitals": "^2.1.4"
