@@ -8,14 +8,6 @@ import { Routes, Route, useLocation } from "react-router-dom";
 import { TransitionGroup, CSSTransition } from "react-transition-group";
 
 import TitleBar from "./components/Titlebar";
-import Home from "./components/Home";
-import MeteorExtinctionHome from "./components/MeteorExtinctionHome";
-import MeteorExtinctionDownload from "./components/MeteorExtinctionDownload";
-import WoodenGUIHome from "./components/WoodenGUIHome";
-import WoodenGUIDownload from "./components/WoodenGUIDownload";
-import Contact from "./components/Contact";
-import IssueTemplate from "./components/IssueTemplate";
-import CheckMEVersion from "./components/CheckMEVersion";
 
 import CustomProgress from "./components/SuspenseFallback";
 
@@ -24,8 +16,8 @@ import {
   setStorageMode,
 } from "./components/services/SettingsService";
 import BrowserTitle from "./components/BrowserTitle";
-import Changelog from "./components/Changelog";
-import NotFound from "./components/404";
+
+import { MyRoutes } from "./Routes";
 
 function App() {
   const [themeMode, setThemeMode] = useState(getStorageMode);
@@ -100,6 +92,7 @@ function App() {
         }
       }
       setLocationBefore(location);
+      setLocationForTransition();
     }
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [location]);
@@ -123,10 +116,11 @@ function App() {
     );
   }, [height]);
 
-  useEffect(() => {
-    setTransitionLocation(location);
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [locationBefore]);
+  function setLocationForTransition() {
+    React.startTransition(() => {
+      setTransitionLocation(location);
+    });
+  }
 
   return (
     <ThemeProvider theme={appliedTheme}>
@@ -145,92 +139,22 @@ function App() {
             onExited={() => setDisplayLocation(location)}
           >
             <Routes location={transitionLocation}>
-              <Route
-                path="*"
-                element={
-                  <div style={{ position: "absolute", width: "100%" }}>
-                    <NotFound />
-                  </div>
-                }
-              ></Route>
-              <Route
-                path="/"
-                element={
-                  <div style={{ position: "absolute", width: "100%" }}>
-                    <Home />
-                  </div>
-                }
-              ></Route>
-              <Route
-                path="meteorextinction"
-                element={
-                  <div style={{ position: "absolute", width: "100%" }}>
-                    <MeteorExtinctionHome themeMode={themeMode} />
-                  </div>
-                }
-              ></Route>
-              <Route
-                path="meteorextinction/download"
-                element={
-                  <div style={{ position: "absolute", width: "100%" }}>
-                    <MeteorExtinctionDownload themeMode={themeMode} />
-                  </div>
-                }
-              ></Route>
-              <Route
-                path="meteorextinction/download/changelog"
-                element={
-                  <div style={{ position: "absolute", width: "100%" }}>
-                    <Changelog
-                      displayLocation={displayLocation}
-                      location={location}
-                    />
-                  </div>
-                }
-              ></Route>
-              <Route
-                path="woodengui"
-                element={
-                  <div style={{ position: "absolute", width: "100%" }}>
-                    <WoodenGUIHome themeMode={themeMode} />
-                  </div>
-                }
-              ></Route>
-              <Route
-                path="woodengui/download"
-                element={
-                  <div style={{ position: "absolute", width: "100%" }}>
-                    <WoodenGUIDownload />
-                  </div>
-                }
-              ></Route>
-              <Route
-                path="contact"
-                element={
-                  <div style={{ position: "absolute", width: "100%" }}>
-                    <Contact themeMode={themeMode} />
-                  </div>
-                }
-              ></Route>
-              <Route
-                path="contact/issue-template"
-                element={
-                  <div style={{ position: "absolute", width: "100%" }}>
-                    <IssueTemplate themeMode={themeMode} />
-                  </div>
-                }
-              ></Route>
-              <Route
-                path="meteorextinction/download/update"
-                element={
-                  <div style={{ position: "absolute", width: "100%" }}>
-                    <CheckMEVersion
-                      displayLocation={displayLocation}
-                      location={location}
-                    />
-                  </div>
-                }
-              ></Route>
+              {MyRoutes.map((route) => {
+                return (
+                  <Route
+                    path={route.path}
+                    key={route.path}
+                    element={
+                      <div style={{ position: "absolute", width: "100%" }}>
+                        <route.element
+                          themeMode={themeMode}
+                          displayLocation={displayLocation}
+                        />
+                      </div>
+                    }
+                  />
+                );
+              })}
             </Routes>
           </CSSTransition>
         </TransitionGroup>
