@@ -3,23 +3,10 @@ import React from "react";
 import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
 import { styled } from "@mui/material/styles";
-import { makeStyles } from "@mui/styles";
+import styles from "./BannerHome.module.css";
 import { AppBar } from "@mui/material";
 
 import LastandDevelopmentLogo from "./icons/LastandDevelopmentLogo.svg";
-
-const useStyles = makeStyles((theme) => ({
-  Logo: {
-    "@media screen and (min-width: 450px)": {
-      height: "120px",
-      width: "307.34px",
-    },
-    "@media screen and (max-width: 450px)": {
-      height: "27vw",
-      width: "69.15vw",
-    },
-  },
-}));
 
 const StyledToolbar = styled(Toolbar)(({ theme }) => ({
   alignItems: "flex-start",
@@ -28,8 +15,6 @@ const StyledToolbar = styled(Toolbar)(({ theme }) => ({
 }));
 
 export default function BannerHome() {
-  const classes = useStyles();
-
   return (
     <Box sx={{ flexGrow: 1 }}>
       <AppBar
@@ -42,7 +27,7 @@ export default function BannerHome() {
           className="mui-fixed"
           sx={{ justifyContent: "center" }}>
           <img
-            className={classes.Logo}
+            className={styles.Logo}
             src={LastandDevelopmentLogo}
             alt="Lastand Development"
             component="div"
