@@ -5,25 +5,8 @@ import MenuItem from "@mui/material/MenuItem";
 import LanguageRoundedIcon from "@mui/icons-material/LanguageRounded";
 import { useTranslation } from "react-i18next";
 import { ListItemIcon, ListItemText } from "@mui/material";
-import { makeStyles } from "@mui/styles";
-
-const useStyles = makeStyles((theme) => ({
-  menuItem: {
-    margin: "4px 6px 4px 6px !important",
-    paddingLeft: "10px !important",
-    paddingRight: "10px !important",
-    borderRadius: "10px !important",
-    transition: "background-color 0.2s ease-in-out",
-  },
-  menu: {
-    paddingTop: "2px !important",
-    paddingBottom: "2px !important",
-  },
-}));
 
 export default function LanguageMenu(props) {
-  const classes = useStyles();
-
   const [t] = useTranslation();
   const { i18n } = useTranslation();
   const [anchorEl, setAnchorEl] = useState(null);
@@ -68,9 +51,7 @@ export default function LanguageMenu(props) {
 
   return (
     <Fragment>
-      <MenuItem
-        onClick={handleClick}
-        className={classes.menuItem}>
+      <MenuItem onClick={handleClick}>
         <ListItemIcon>
           <LanguageRoundedIcon fontSize="medium" />
         </ListItemIcon>
@@ -82,7 +63,6 @@ export default function LanguageMenu(props) {
             width: 150,
           },
         }}
-        classes={{ list: classes.menu }}
         id="language-menu"
         elevation={3}
         anchorEl={anchorEl}
@@ -101,7 +81,6 @@ export default function LanguageMenu(props) {
           <MenuItem
             key={"langItem" + i}
             selected={i === selectedIndex}
-            className={classes.menuItem}
             onClick={() => {
               i18n.changeLanguage(l.lang);
               handleClose();
