@@ -22,7 +22,7 @@ import { ReactComponent as GitHub } from "./icons/GitHubIcon.svg";
 
 import { useTranslation } from "react-i18next";
 
-import { makeStyles } from "@mui/styles";
+import styles from "./MeteorExtinctionHome.module.css";
 
 import { useSwipeable } from "react-swipeable";
 
@@ -47,57 +47,6 @@ import rocketexplosionM from "./imagesME/rocketexplosion/M.webp";
 import rocketexplosionS from "./imagesME/rocketexplosion/S.webp";
 
 import { Helmet, HelmetProvider } from "react-helmet-async";
-
-const useStyles = makeStyles((theme) => ({
-  ImageCarouselBox: {
-    display: "flex",
-    justifyContent: "center",
-    alignItems: "center",
-  },
-  ImageCarouselContent: {
-    width: "100%",
-    height: "100%",
-    borderRadius: "10px",
-    position: "absolute",
-    top: 0,
-    left: 0,
-  },
-  ImageCarouselContentBox: {
-    width: "100%",
-    aspectRatio: "16 / 9",
-    overflow: "hidden",
-    borderRadius: "10px",
-    position: "relative",
-  },
-  ImageCarouselIconNext: {
-    "@media screen and (min-width: 1200px)": {
-      fontSize: "40px !important",
-    },
-    "@media screen and (max-width: 1200px)": {
-      fontSize: "4vw !important",
-    },
-  },
-  ImageCarouselIconBefore: {
-    "@media screen and (min-width: 1200px)": {
-      fontSize: "40px !important",
-    },
-    "@media screen and (max-width: 1200px)": {
-      fontSize: "4vw !important",
-    },
-  },
-  PointsDiv: {
-    "@media screen and (min-width: 1200px)": {
-      marginTop: "12px",
-    },
-    "@media screen and (max-width: 1200px)": {
-      marginTop: "1vw",
-    },
-    textAlign: "center",
-    height: "10px",
-    display: "flex",
-    justifyContent: "center",
-  },
-}));
 
 const Dot = styled("div")(({ theme, selected }) => ({
   "@media screen and (min-width: 1200px)": {
@@ -129,8 +78,6 @@ const Dot = styled("div")(({ theme, selected }) => ({
 }));
 
 export default function MeteorExtinctionHome(props) {
-  const classes = useStyles();
-
   const [t] = useTranslation();
 
   const [imageIndex, setImageIndex] = useState(1);
@@ -234,7 +181,7 @@ export default function MeteorExtinctionHome(props) {
                 padding: "3%",
               }}
             >
-              <Box className={classes.ImageCarouselBox}>
+              <Box className={styles.ImageCarouselBox}>
                 <Box
                   sx={{
                     marginRight: "1.5%",
@@ -253,14 +200,14 @@ export default function MeteorExtinctionHome(props) {
                     onClick={previousImage}
                   >
                     <NavigateBeforeRounded
-                      className={classes.ImageCarouselIconBefore}
+                      className={styles.ImageCarouselIconBefore}
                     />
                   </IconButton>
                 </Box>
                 <Box sx={{ width: "87%", height: "100%" }}>
                   <Box
                     {...handlers}
-                    className={classes.ImageCarouselContentBox}
+                    className={styles.ImageCarouselContentBox}
                     flexGrow={1}
                   >
                     <Slide
@@ -281,7 +228,7 @@ export default function MeteorExtinctionHome(props) {
                       }
                     >
                       <img
-                        className={classes.ImageCarouselContent}
+                        className={styles.ImageCarouselContent}
                         src={imageNames[0]}
                         alt="Gameplay"
                       />
@@ -304,7 +251,7 @@ export default function MeteorExtinctionHome(props) {
                       }
                     >
                       <img
-                        className={classes.ImageCarouselContent}
+                        className={styles.ImageCarouselContent}
                         src={imageNames[1]}
                         alt={t("meteorExtinctionImages.meteorDestroyed")}
                       />
@@ -327,7 +274,7 @@ export default function MeteorExtinctionHome(props) {
                       }
                     >
                       <img
-                        className={classes.ImageCarouselContent}
+                        className={styles.ImageCarouselContent}
                         src={imageNames[2]}
                         alt={t("meteorExtinctionImages.meteorExplosion")}
                       />
@@ -350,7 +297,7 @@ export default function MeteorExtinctionHome(props) {
                       }
                     >
                       <img
-                        className={classes.ImageCarouselContent}
+                        className={styles.ImageCarouselContent}
                         src={imageNames[3]}
                         alt={t("meteorExtinctionImages.rocket")}
                       />
@@ -373,13 +320,13 @@ export default function MeteorExtinctionHome(props) {
                       }
                     >
                       <img
-                        className={classes.ImageCarouselContent}
+                        className={styles.ImageCarouselContent}
                         src={imageNames[4]}
                         alt={t("meteorExtinctionImages.rocketExploding")}
                       />
                     </Slide>
                   </Box>
-                  <div className={classes.PointsDiv}>
+                  <div className={styles.PointsDiv}>
                     <Dot
                       theme={props.themeMode}
                       selected={imageIndex === 1} />
@@ -411,7 +358,7 @@ export default function MeteorExtinctionHome(props) {
                     onClick={nextImage}
                   >
                     <NavigateNextRounded
-                      className={classes.ImageCarouselIconNext}
+                      className={styles.ImageCarouselIconNext}
                     />
                   </IconButton>
                 </Box>
